Rename Picture to PLTPicture and remove unnecessary methods
